Assignment6

+++ server/controllers/fileController.js
@@ -1,16 +1,25 @@
-
-const Bill = require('../models').Bill;
-const User = require('../models').User;
-const File = require('../models').File;
+const Bill = require("../models/bill").Bill;
+const File = require("../models/fileModel").Files;
+const User = require('../models/userModel').User;
 const moment = require('moment');
 const fs = require('fs');
 const md5File = require('md5-file');
 
 moment.suppressDeprecationWarnings = true;
 
+const uuidv4 = require('uuid/v4');
 const path = require('path');
 
 const multer = require('multer');
+// const bucket = "to run locally uncomment this!";
+const aws = require('aws-sdk');
+const s3 = new aws.S3({ apiVersion: '2006-03-01' });
+aws.config.update({
+    region: 'us-east-1'
+});
+
+const multerS3 = require('multer-s3');
+const bucket = process.env.S3_BUCKET;
 const storage = multer.diskStorage({
     destination: (req, file, cb) => {
         cb(null, 'public/uploads')
@@ -19,6 +28,28 @@ const storage = multer.diskStorage({
         cb(null, Date.now() + '-' + file.originalname)
     }
 });
+const uploadS3 = multer({
+    fileFilter: function (req, file, callback) {
+        var ext = path.extname(file.originalname);
+        if (ext !== '.png' && ext !== '.jpg' && ext !== '.pdf' && ext !== '.jpeg') {
+            return callback({ "Error": "Only pdfs & images are allowed" }, false);
+        }
+        callback(null, true)
+    },
+    limits: {
+        fileSize: 1024 * 1024
+    },
+    storage: multerS3({
+        s3: s3,
+        bucket: bucket,
+        acl: 'private',
+        contentType: multerS3.AUTO_CONTENT_TYPE,
+        //   serverSideEncryption: 'AES256',
+        key: function (req, file, cb) {
+            cb(null, Date.now() + '-' + file.originalname)
+        }
+    })
+}).single('billAttachment');
 
 const upload = multer({
     storage: storage,
@@ -77,48 +108,46 @@ module.exports = {
                     }
                     else {
 
-                        upload(req, res, function (err) {
+                        uploadS3(req, res, function (err) {
                             if (err) {
                                 return res.status(400).send(err);
                             } else {
-                                md5File(`public/uploads/${req.file.filename}`, (err10, hash) => {
-                                    if (err10) {
+                                return File
+                                    .create({
+                                        id: uuidv4(),
+                                        file_name: req.file.originalname,
+                                        url: req.file.location,
+                                        upload_date: new Date(),
+                                        size: req.file.size,
+                                        fileOwner: user.dataValues.email_address,
+                                        bill: bills[0].dataValues.id,
+                                        md5: "fieldName =>" + req.file.fieldname,
+                                        key: req.file.key
+                                    })
+                                    .then((file) => {
+                                        delete file.dataValues.createdAt;
+                                        delete file.dataValues.updatedAt;
+                                        delete file.dataValues.fileOwner;
+                                        delete file.dataValues.size;
+                                        delete file.dataValues.bill;
+                                        delete file.dataValues.key;
+                                        delete file.dataValues.md5;
+                                        Bill
+                                            .update(
+                                                { attachment: file.dataValues.id },
+                                                {
+                                                    where: {
+                                                        id: req.params.id
+                                                    }
+                                                }
+                                            )
+                                        res.status(201).send(file);
+                                    })
+                                    .catch((error) => {
                                         res.status(400).send(error);
-                                    }
-                                    else {
-                                        return File
-                                            .create({
-                                                file_name: req.file.filename,
-                                                url: `public/uploads/${req.file.filename}`,
-                                                upload_date: new Date(),
-                                                size: req.file.size,
-                                                fileOwner: user.dataValues.email_address,
-                                                bill: bills[0].dataValues.id,
-                                                md5: hash
-                                            })
-                                            .then((file) => {
-                                                delete file.dataValues.createdAt;
-                                                delete file.dataValues.updatedAt;
-                                                delete file.dataValues.fileOwner;
-                                                delete file.dataValues.size;
-                                                delete file.dataValues.bill;
-                                                delete file.dataValues.md5;
-                                                Bill
-                                                    .update(
-                                                        { attachment: file.dataValues.id },
-                                                        {
-                                                            where: {
-                                                                id: req.params.id
-                                                            }
-                                                        }
-                                                    )
-                                                res.status(201).send(file);
-                                            })
-                                            .catch((error) => {
-                                                res.status(400).send(error);
-                                            });
-                                    }
-                                })
+                                    });
+
+
                             }
                         });
                     }
@@ -193,6 +222,7 @@ module.exports = {
                                 delete file[0].dataValues.updatedAt;
                                 delete file[0].dataValues.fileOwner;
                                 delete file[0].dataValues.size;
+                                delete file[0].dataValues.key;
                                 delete file[0].dataValues.bill;
                                 delete file[0].dataValues.md5;
                                 res.status(200).send(file[0]);
@@ -271,7 +301,7 @@ module.exports = {
                                         }
                                     })
                                     .then((file) => {
-                                        
+
                                         if (file.length == 0) {
                                             return res.status(404).send({
                                                 message: "File Not Found!"
@@ -282,23 +312,32 @@ module.exports = {
                                                 message: "File for this Bill Not Found!"
                                             })
                                         }
-                                        fs.unlink(file[0].dataValues.url, function (err) {
-                                            
-                                        return File
-                                        
-                                            .destroy({
-                                                where: {
-                                                    id: req.params.fileId
-                                                }
-                                            })
-                                            .then((rowDeleted) => {
-                                                if (rowDeleted === 1) {
-                                                    res.status(204).send('Deleted successfully');
-                                                }
-                                            })
-                                            .catch((error2) => {
-                                                res.status(400).send(error2);
-                                            });
+                                        s3.deleteObject({
+                                            Bucket: bucket,
+                                            Key: file[0].key
+                                        }, function (err09) {
+                                            if (err09) {
+                                                return res.status(400).send({
+                                                    message: "Error while deleting from S3!"
+                                                })
+                                            } else {
+
+                                                return File
+
+                                                    .destroy({
+                                                        where: {
+                                                            id: req.params.fileId
+                                                        }
+                                                    })
+                                                    .then((rowDeleted) => {
+                                                        if (rowDeleted === 1) {
+                                                            res.status(204).send('Deleted successfully');
+                                                        }
+                                                    })
+                                                    .catch((error2) => {
+                                                        res.status(400).send(error2);
+                                                    });
+                                            }
                                         })
                                     })
                                     .catch((error) => {
